Refactor trip planner components: update search bar layout, add trip button, and enhance places repository with batch insert and radius optimization

+++ lib/src/features/trip_planner/data/datasource/place_remote_data_source.dart
@@ -37,7 +37,7 @@ class PlaceRemoteDataSource implements PlaceDataSource {
                 'latitude': lat,
                 'longitude': lng,
               },
-              'radius': radius.toInt(),
+              'radius': radius,
             },
           },
           if (categories != null && categories.isNotEmpty)
